chore: GuestbookSearchResponse의 rating을 실수로 변경

+++ src/main/java/com/babzip/backend/guestbook/dto/response/GuestbookSearchResponse.java
@@ -8,7 +8,7 @@ public record GuestbookSearchResponse (
         String restaurantName,
         String address,
         String content,
-        Long rating,
+        Double rating,
         LocalDateTime createdAt
 ){
     public static GuestbookSearchResponse toDto(Guestbook guestbook) {
